Change icon position
Change drawing to make better use of top spacings. Icons are shifted
upwards by modifying the drawing Rectangle, then moved down by adjusting
the alignment. This is kind of a hack, but then again changing the
vertical span to change icon position and save on space was already kind
of a hack. Eventually hopefully a similiar positioning can be achieved a
bit nicer.