Fix object-based parameters not always accessible
String values in Parameters that are based on objects, such as
"display-nick", couldn't be accessed outside of actual replacement in
the Identifier class. This broke the reply command, since it tries to
access user-based information from the given Parameters object.

The fix moves getting user-based keys from the Identifier to the
Parameters class, so they can always be accessed.