Notification: support matching several channels
If you want to set up a notification which triggers in several channels,
you have to create multiple notifications with different `channel`
setting, which is tedious.

Allow using a comma-separated list of channels in NotificationEditor, so
that users may be able to set their notifications for multiple channels
at the same time.

Impelementation details
-----------------------
Use existing `Helper.parseChannelsFromString` for parsing of the
serialized setting.  Case-insensitive TreeSet in class Notification is
needed to preserve the existing behavior of method `matchesChannel`.

Older versions of Chatty save a single String for this setting, which is
a channel name. A single channel name is still a comma-separated list of
channel names, consisting of a single item (with no commas).  So no
migration is needed.