Support matching several channels in notifications (#428)
* Helper: add tests for parseChannelsFromString & buildStreamsString

* Notification: drop unused field soundFileDelay

Field soundFileDelay hasn't been used since class Notification was
introduced in commit 3d789d4a (Add new Notification System, 2017-05-16).

* Notification: add test for matchesChannel

* Notification: support matching several channels

If you want to set up a notification which triggers in several channels,
you have to create multiple notifications with different `channel`
setting, which is tedious.

Allow using a comma-separated list of channels in NotificationEditor, so
that users may be able to set their notifications for multiple channels
at the same time.

Impelementation details
-----------------------
Use existing `Helper.parseChannelsFromString` for parsing of the
serialized setting.  Case-insensitive TreeSet in class Notification is
needed to preserve the existing behavior of method `matchesChannel`.

Older versions of Chatty save a single String for this setting, which is
a channel name. A single channel name is still a comma-separated list of
channel names, consisting of a single item (with no commas).  So no
migration is needed.

* NotificationTest: check that channels not in the list don't match

Previously added tests don't check cases when `matchesChannel()` is
supposed to return `false`.  Add negative tests to NotificationTest to
make sure that the tests fail in case matchesChannel() would always
return true by accident.

Note, that for testThatMatchingAgainstNullChannelsWorks(), any channel
matches a notification with channels==null.