UpdateDialog: show links to help-releases.html (#411)
* MainGui: use switch statement on link type in MyLinkLabelListener

This refactoring has been made automatically using IntelliJ IDEA's
inspection "'if' replaceable with 'switch'".
https://www.jetbrains.com/help/idea/list-of-java-inspections.html#java-language-level-migration-aids

* HTML: fix invalid padding syntax in CSS

Non-zero values for padding in CSS must have the unit specified.  So add
unit `px` where it is missing.

* UpdateDialog: print clicked link type for debugging

Standalone UpdateDialog prints clicked links to System.out.  Print
`type` in addition to `ref` for debugging.

* UpdateDialog: comment out unused GridBagConstraints

Because JLabel for BETA_INFO is commented out, this instance of
GridBagConstraints is unused.  So comment it out as well to avoid
confusion.

* UpdateDialog: don't place &nbsp; inside HTML syntax

`replaceAll` which changes all spaces to HTML entity `&nbsp;` is called
on a big HTML String, which causes the variable `main` to include

    <div&nbsp;style=...

as part of its content.  So spell out &nbsp; manually only in the
human-readable part of the HTML string.

* UpdateDialog: show links to help-releases.html

User who clicks on the "Check for update" menu item might be interested
in release notes, so add a couple of links to the file help-releases.html:

  1. at the bottom of UpdateDialog, always shown;
  2. as words "latest version" in the text "You are running the latest
     version" in the UpdateDialog, shown only when the latest version
     is being used.

Note, that `y` coord in GridBagConstraints for `closeButton` just below
does not need to be updated, since its already uses value `6`.