Emote modifiers fix
Check for emote modifiers in emote id rather than assuming that all
non-numeric emote ids are due to a modified emote (as it was at some
point). The emote info stuff should be updated further at some point,
since it probably still contains a lot of stuff that only made sense in
older versions of the code/APIs.