Make settings file follow xdg specification
This commit changes `getUserDataDirectory` to return `$XDG_CONFIG_HOME/chatty`, or `$HOME/.config/chatty` if `$XDG_CONFIG_HOME` doesn't exist.
In addition, this commit will cause Chatty to, on startup, detect if the legacy settings directory exists ($HOME/.chatty) and, if it does, move it to whatever `getUserDataDirectory` returns. It does this on startup such that the new settings directory can be taken into effect without needing to restart the client, making for a seamless transition.