Popout attributes changes
Prevent duplicates and limit number of saved entries. This is especially
because popouts opened by drag&drop or loading layouts don't remove an
popoutAttributes entry, but can still add an entry to it. This can lead
to a huge list of sometimes even identical entries.

Prevent a popout being opened using attributes that are equal to an
already open popout. Open popouts attributes are also saved when
settings are written, which can make sense if the layout is not
restored, but otherwise probably not.