Fix popouts getting the same id sometimes
Issue: A layout that contains two popouts with the same id will cause
two popouts to open, but only one will contain all tabs from both
popouts. The following are two cases how a layout like that could
happen.

Reused popouts:
* Open f0
* Open f1
* Close f0
* Load layout with f1, which closes f1 (all popouts), but reuses f0
  (take first unused) and renames it to f1 so that content is added to
  the correct popout
* Open another popout, which reuses f1, now there are two popouts named
  f1

Due to using a counter to set the default id:
* Load a layout (e.g. on start) with f1, popout f0 gets created
  (increases counter), renamed to f1 so that content is added to the
  correct popout
* Open another popout, popout f1 gets created (since the counter is at
  1), now there are two popouts named f1

This is fixed by making sure that a created or reused popout will
default to an id that doesn't clash with any currently used ones.