Add settings to toggle error sound
The provideErrorFeedback method of the LookAndFeel seems to play a
system sound by default which is triggered in some cases, mosty commonly
when backspacing in a text input field when backspacing isn't possible.

To disable the sound, the specific LookAndFeel classes are overridden to
disable the method. For JTattoo an option is added, so that not each
individual LookAndFeel implementation has to be changed. For LookAndFeel
where the class is not clear beforehand it's not implemented right now,
since that seems a bit more complicated.