Fix emote size bug, refactor for clarity
Since BTTV emotes don't have the size in the API it relies more on
getting the size from the image, which in case of loading animated WebP
images did not correct for the size factor. That size was then saved in
the Emoticon object, so changing emote scaling afterwards would not
change the initial behaviour (either correct or wrong base size,
depending on whether a 1x or 2x image was loaded first). Emote size
caching to file should only affect the temp loading placeholder image.