assets: Add linux specific assets
This will be used by distro packagers to help integrate chatty into a linux install.

1. Provide a .desktop file
    - https://specifications.freedesktop.org/desktop-entry-spec/latest/
    - Distros need to install this to:
      `/usr/share/applications/io.github.chatty.desktop`
    - Distros may provide a wrapper script such as:
      `exec $JAVA_HOME/bin/java -jar /usr/share/chatty/Chatty.jar "$@"`
      installed to `/usr/bin/chatty`.
    - Validated with `desktop-file-validate`

2. Provide appstream metainfo file
    - https://www.freedesktop.org/wiki/Distributions/AppStream/
    - Distros need to install this to:
      `/usr/share/metainfo/io.github.chatty.metainfo.xml`
    - Validated with `appstreamcli validate --pedantic`

The metainfo will need to be updated if any URLs change for example, but,
otherwise they will require no maintenance.

One nice to have in the future would be to autogenerate the `<releases>`,
using `appstreamcli news-to-metainfo` from a generic text file, however,
this isn't neccessary.